requirements.txt
matplotlib
nibabel
pandas
scipy
tqdm
sklearn
pyyaml
